MSI: 388: Adapt `\Magento\CatalogInventory\Model\ResourceModel\Stock\Status::addStockStatusToSelect
