+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -38,7 +38,9 @@
                 type="Magento\InventoryCatalog\Plugin\CatalogInventory\UpdateSourceItemAtLegacyStockItemSavePlugin"/>
     </type>
     <type name="Magento\CatalogInventory\Model\ResourceModel\Stock\Status">
-        <plugin name="adapt_resource_model_stock_status_to_multi_stocks"
-                type="Magento\InventoryCatalog\Plugin\CatalogInventory\AdaptResourceModelStockStatusToMultiStocks"/>
+        <plugin name="adapt_add_stock_data_to_collection_to_multi_stocks"
+                type="Magento\InventoryCatalog\Plugin\CatalogInventory\Model\ResourceModel\Stock\Status\AdaptAddStockDataToCollectionToMultiStocks"/>
+        <plugin name="adapt_add_stock_status_to_select_to_multi_stocks"
+                type="Magento\InventoryCatalog\Plugin\CatalogInventory\Model\ResourceModel\Stock\Status\AdaptAddStockStatusToSelectToMultiStocks"/>
     </type>
 </config>
